+++ src/component/Modal/TaskForm.tsx
@@ -1,28 +1,13 @@
-import { useState } from "react";
 import { v4 as uuidv4 } from "uuid";
+import { useData } from "../../context/dataContext";
 import "./TaskForm.css";
 
-type FormData = {
-  title: string;
-  description: string;
-  time: string;
-};
-
-type ListOfTasks = { tasks: any[] };
-
-const initialFormData: FormData = {
-  title: "",
-  description: "",
-  time: "",
-};
-
 const TaskForm = () => {
-  const [formData, setFormData] = useState<FormData>(initialFormData);
-  const [listOfTasks, setListOfTasks] = useState<ListOfTasks>({ tasks: [] });
+  const { formData, setListOfTasks, setFormData } = useData();
 
   const handleTaskDetail = (event: any) => {
     event.preventDefault();
-    setListOfTasks((prev) => {
+    setListOfTasks((prev: any) => {
       const updatedTasks = prev.tasks.concat({
         id: uuidv4(),
         title: formData.title,
@@ -52,7 +37,7 @@ const TaskForm = () => {
                 placeholder="Add Title"
                 className="input-title"
                 onChange={(event) =>
-                  setFormData((prev) => ({
+                  setFormData((prev: any) => ({
                     ...prev,
                     title: event.target.value,
                   }))
@@ -66,7 +51,7 @@ const TaskForm = () => {
                 placeholder="Add Description"
                 className="input-description"
                 onChange={(event) =>
-                  setFormData((prev) => ({
+                  setFormData((prev: any) => ({
                     ...prev,
                     description: event.target.value,
                   }))
@@ -81,7 +66,7 @@ const TaskForm = () => {
                 placeholder="Add Time"
                 className="input-title"
                 onChange={(event) =>
-                  setFormData((prev) => ({
+                  setFormData((prev: any) => ({
                     ...prev,
                     time: event.target.value,
                   }))
